rename : common패키지 생성 및 파일이동, 주석 추가
