+++ BE/girin-grim/src/main/java/com/example/giringrim/utils/common/ApiResponseGenerator.java
@@ -1,5 +1,6 @@
-package com.example.giringrim.utils;
+package com.example.giringrim.utils.common;
 
+import com.example.giringrim.utils.common.ApiResponse;
 import com.example.giringrim.utils.exception.Error;
 import lombok.experimental.UtilityClass;
 import org.springframework.http.HttpStatus;
